+++ src/SWP/Bundle/CoreBundle/Theme/Installer/TenantAwareThemeInstaller.php
@@ -18,6 +18,7 @@ namespace SWP\Bundle\CoreBundle\Theme\Installer;
 
 use SWP\Bundle\CoreBundle\Twig\Cache\TenantAwareCacheInterface;
 use SWP\Component\MultiTenancy\Context\TenantContextInterface;
+use Sylius\Bundle\ThemeBundle\Asset\Installer\AssetsInstallerInterface;
 use Sylius\Bundle\ThemeBundle\Loader\ThemeLoaderInterface;
 use SWP\Bundle\CoreBundle\Theme\Model\ThemeInterface;
 use Symfony\Component\Filesystem\Filesystem;
@@ -49,19 +50,39 @@ final class TenantAwareThemeInstaller implements ThemeInstallerInterface
     private $baseDir;
 
     /**
+     * @var AssetsInstallerInterface
+     */
+    private $assetsInstaller;
+
+    /**
+     * @var string
+     */
+    private $assetsDir;
+
+    /**
      * TenantAwareThemeInstaller constructor.
      *
-     * @param TenantContextInterface $tenantContext
-     * @param ThemeLoaderInterface   $themeLoader
-     * @param \Twig_Environment      $twig
-     * @param string                 $baseDir
+     * @param TenantContextInterface   $tenantContext
+     * @param ThemeLoaderInterface     $themeLoader
+     * @param \Twig_Environment        $twig
+     * @param string                   $baseDir
+     * @param AssetsInstallerInterface $assetsInstaller
+     * @param string                   $assetsDir
      */
-    public function __construct(TenantContextInterface $tenantContext, ThemeLoaderInterface $themeLoader, \Twig_Environment $twig, string $baseDir)
-    {
+    public function __construct(
+        TenantContextInterface $tenantContext,
+        ThemeLoaderInterface $themeLoader,
+        \Twig_Environment $twig,
+        string $baseDir,
+        AssetsInstallerInterface $assetsInstaller,
+        string $assetsDir
+    ) {
         $this->tenantContext = $tenantContext;
         $this->themeLoader = $themeLoader;
         $this->twig = $twig;
         $this->baseDir = $baseDir;
+        $this->assetsInstaller = $assetsInstaller;
+        $this->assetsDir = $assetsDir;
     }
 
     /**
@@ -98,6 +119,8 @@ final class TenantAwareThemeInstaller implements ThemeInstallerInterface
         if ($cache instanceof TenantAwareCacheInterface) {
             $filesystem->remove($cache->generateCacheDir());
         }
+
+        $this->assetsInstaller->installAssets($this->assetsDir, AssetsInstallerInterface::HARD_COPY);
 
         return $theme;
     }