+++ CHANGELOG.md
@@ -1,13 +1,17 @@
 
-CHANGELOG for version <= 1.3.x
+CHANGELOG for version <= 1.4.x
 ==============================
 
 This changelog references the relevant changes (bug and security fixes) done in <= 1.1 minor versions.
 
 To get the diff for a specific change, go to https://github.com/superdesk/web-publisher/commit/XXX where XXX is the change hash
 
-* 1.3.0
+* 1.4.0
  * feature [#459] Added an option to preview an article before even publishing it
+
+* 1.3.0
+ * feature [#462] Added template widget
+ * bug [#460] Install assets when theme is installed
  * feature [#458] Added redirect and notFound functions for twig templates
  * feature [#454] Override existing rules by publish destinations
  * improvement [#455] Added slug to article author - allow to load authors by slug